Change design to use a memory allocator
Cuda doesnt allow for cudaMalloc in the middle of a kernel,
thus I changed the design to grab a bunch of RAM at the beginning.
This memory can then be split amount file requests. Additionally
there is a bit of refactoring. I want to create a unified interface
that allows for importing only one header for all of gvm fs.